renamed schema file
